CRF-4: Separated out housekeeping integration tests
Moved housekeeping integration tests from JudgmentRepositoryIntTest to new JudgmentRepositoryHousekeepingIntTest class.
